Use group_service fixture in all search tests
Enable the group_service fixture automatically for all search tests. A
lot of search tests need this fixture, so avoid them all having to
enable it one by one.
